Move some functions from lib to rc-ui-utils
